New module kas-core::event::cx::timer
